Modify Dockerfile for TeX Live minimal install
Using Debian packages makes it difficult to ensure that LaTeX packages
are up-to-date. For example, stretch (currently latest for Debian's
Docker image) uses TeX Live 2016 as its base, and many packages have
not been updated since 2016! Thus, it would be ideal to have an image
that contains TeX Live and uses tlmgr to install LaTeX packages --
just those that are required -- and to keep those packages up-to-date.

Resolves #2